% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hfspace_shiny_dockerfile.R
\name{generate_hfspace_shiny_dockerfile}
\alias{generate_hfspace_shiny_dockerfile}
\title{Generate Shiny Dockerfile for Hugging Face}
\usage{
generate_hfspace_shiny_dockerfile(
  base = "rocker/r2u:22.04",
  cran_pkgs_name,
  gh_pkgs_repo = NULL,
  port = 7860
)
}
\arguments{
\item{base}{base image}

\item{cran_pkgs_name}{vector of package names required by your Shiny app}

\item{gh_pkgs_repo}{vector of dev package hosted in GitHub required by your Shiny app}

\item{port}{exposed port in dockerfile}
}
\description{
Generate Shiny Dockerfile for Hugging Face
}
