% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hfspace_shiny_iframe.R
\name{generate_hfspace_shiny_iframe}
\alias{generate_hfspace_shiny_iframe}
\title{Generate index.html to embedd Shiny app deploted at Hugging Face}
\usage{
generate_hfspace_shiny_iframe(username, space, webtitle = "It's so shiny~")
}
\arguments{
\item{username}{username at Hugging Face}

\item{space}{name of space at Hugging Face}

\item{webtitle}{title of page (not a heading!)}
}
\description{
Generate index.html to embedd Shiny app deploted at Hugging Face
}
